Fix a problem with login not updating correctly
The login was not getting updated correctly because it would no longer
update if it already loaded the profiles once. Removing this check but
keeping the loading dependency will cause an infinite loop. So we need
to remove this dependency but still check for it to not start a double
request for updating the profiles if another is already in progress.