RNP-91: Move AuthButton -> AuthContainer + change functionality
The authContainer is responsible for handling the user authentication
and the selection of the userprofile/authentictor. There is now a switch
for deciding if you want to use preffered authenticator or not. And
modals will show in the case that you do not want to use the preferred
authentication method.