RNP-109: Rework ChangeAuthView and use generic registerAuthenticator
This commit makes the ChangeAuthView a lot easier to read by not passing
lots of state setters into a helper everywhere. This now only happens in
1 place which is properly named.
The generic helpers have now also been moved to a AuthenticationHelper
instead of being in FingerprintHelper and some of the redundant helper
functions have been removed.