RNP-109: Make updating of the authenticators more robust
Previously if you would encounter an error the ui would not update
properly. With this commit we now update the authenticators when we
encounter en error during the enabling/disabling of fingerprint and
when setting the preferred authenticator. The retrieving of the
authenticators is now also properly error handled if any errors occur
during that process.