RNP-86: Update colors, buttons, and switches with new styling
This commit replaces all old colors with definitons in the AppColors
file.
Additionally it replaces the Buttons and switches with those from the
ReactNative paper package. These are styled using the theme definition
in Theme.ts which is accessed using the provider in AppBootstrap.tsx

It also contains some other minor styling changes, and removes the old
Button.tsx file.